requirements.txt
annotated-types
anyio
asyncio
asyncpg
bcrypt==3.2.2
click
colorama
dnspython
ecdsa
email_validator
fastapi
greenlet
h11
idna
passlib
pyasn1
pydantic
pydantic-settings
pydantic_core
python-dotenv
python-jose
python-multipart
rsa
six
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
